Split writing flow
